requirements/dev.txt
#
# This file is autogenerated by pyproject2conda
# with the following command:
#
#     $ pyproject2conda project --verbose
#
# You should not manually edit this file.
# Instead edit the corresponding pyproject.toml file.
#
dask;python_version=="3.11"
docstring-inheritance
ipykernel
ipython
lazy-loader
module-utilities>=0.6
mypy>=1.11.2
nbclient>=0.10.0
nbval
nox>=2024.4.15
numba>=0.58
numpy>1.24
packaging
pandas-stubs
pytest
pytest-accept
pytest-cov
pytest-sugar
pytest-xdist
scipy;python_version=="3.11"
typing-extensions
xarray>=2023.9.0
